Add files via upload
The project starts with the analysis of the existing data, loading the csv, and deleting the unwanted columns. A word cloud is generated with the help of Generate_wordclouds.py for the tweets after combining all the tweets into one sentence.
Next, the Clean_Train.py function has to be run. Then the function message cleaning is applied to the tweets from file Pipeline.py which removes the punctuation and stop words from the tweet. This gives us the clean data ready for analysis and training. The tweets are then tokenized and the data is split in the 80:20 ratio for testing our Naive Bayes Classifier.

The Test.py is used to test our Naive Bayes Classifier for 20% of the data the model hasn't been exposed to yet. The confusion matrix and classification report are obtained then. To improve the accuracy of the model various other classification techniques such as Logistic Regression / XGBoost etc can be used. Because the data is heavily skewed towards the positive sentiment samples, the negative sentiment samples can be upsampled and the model can be trained in order to improve accuracy.